Feat/normalize numbers to big decimal (#21)
* Changed all big ints to big decimals

* All swap values use token decimals

* Added Orderbook contracts

* Using decimalize in all converter mappings

* Feature/add volume and txCount to WRBTC/usd candlesticks (#22)

* changed mainnet log level to info

* add volume + txCount to WRBTC/USD sticks

* small fix

* removed a useless log msg

* Extra comments

* Change BigInts to Ints where appropriate

* Removed unneccessary comment

+++ src/utils/LendingPool.ts
@@ -1,15 +1,15 @@
 import { SetLoanPool } from '../../generated/ISovryn/ISovryn'
 import { LendingPool } from '../../generated/schema'
-import { BigInt } from '@graphprotocol/graph-ts'
+import { BigDecimal, BigInt } from '@graphprotocol/graph-ts'
 
 export function createAndReturnLendingPool(event: SetLoanPool): LendingPool {
   let lendingPoolEntity = LendingPool.load(event.params.loanPool.toHexString())
   if (lendingPoolEntity == null) {
     lendingPoolEntity = new LendingPool(event.params.loanPool.toHexString())
     lendingPoolEntity.underlyingAsset = event.params.underlying.toHexString()
-    lendingPoolEntity.poolTokenBalance = BigInt.zero()
-    lendingPoolEntity.assetBalance = BigInt.zero()
-    lendingPoolEntity.totalAssetLent = BigInt.zero()
+    lendingPoolEntity.poolTokenBalance = BigDecimal.zero()
+    lendingPoolEntity.assetBalance = BigDecimal.zero()
+    lendingPoolEntity.totalAssetLent = BigDecimal.zero()
     lendingPoolEntity.save()
   }
   return lendingPoolEntity
